+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## Unreleased
 
 **Features**:
+
 - Add configuration to allow high cardinality tags in metrics. ([#4805](https://github.com/getsentry/relay/pull/4805))
 - Make client sdk mandatory for profile sample v2. ([#4853](https://github.com/getsentry/relay/pull/4853))
 - Sharding into many topics based on partition key. ([#4861](https://github.com/getsentry/relay/pull/4861))
 - Parse Chromium stability report from minidumps into context ([#4837](https://github.com/getsentry/relay/pull/4837))
+- Add mechanism to allow ingestion only from trusted relays. ([#4772](https://github.com/getsentry/relay/pull/4772))
 
 **Internal**: